A type may name a run it does not own, count with 'usize', and say what it does as it ends

+++ src/constant.h
@@ -18,14 +18,14 @@ typedef struct {
     const Type *type;
 
     union {
-        int32_t as_int;
+        int64_t as_int;
         float as_float;
         bool as_bool;
         String *as_string;
     };
 } Constant;
 
-static inline Constant constant_int(const Type *type, int32_t value) {
+static inline Constant constant_int(const Type *type, int64_t value) {
     return (Constant){.type = type, .as_int = value};
 }
